_drafts: Add Markdown style guide and abbreviations
Create a new Jekyll supported ‘_drafts’ directory [1]. It may the right
place for not ready-to-publish/must-be-reviewed and outdated files – as
long as it will not be misused as a trash folder.

New files added:
  • _drafts/md_abbreviations.md
  • _drafts/md_style-guide.md

The former file holds a collection of acronym/abbreviation definitions
and is intended to be used as copy ’n paste source for all new Markdown
files. When strictly updated and applied, it will ensure a consistent
appearance for website balloon tips of abbreviations across all pages.

The latter is an attempt establishing useful suggestions for writing
Markdown files for the website.

[1] <https://jekyllrb.com/docs/drafts/>