Upgrade to boostrap 5
The previous header had issues at specific screen sizes. On such issue
is that when the window was somewhat narrow (half the screen width at
21:9), the header would overlap with the top of the regular page
content.

Before setting out to patch details on the header, I though it best to
upgrade to a newer bootstrap version. I then fixed the original issue,
and also improved the vertical alignment of the elements in the header
(which is easier with a newer bootstrap). I also shrunk the Agola search
bar a tiny bit, so that the header does not overflow as easily on
not-too-wide screens.

I'm still not convinced about the GitHub image on the upper right
corner. On small screens, after expanding the header, this image covers
up the menu button, so it's impossible to click it or to collapse the
menu again.